Wave 32: 2641 -> 2785 (+144), bytes 26.63% -> 29.26%
Three agents on address-locality blocks, this time at a 76-160 byte ceiling.
Every function verified by trymatch exit status in sweeps run after each agent
went idle; both builds reproduce 14dd0b22c894865867aff89e8116b2dffae25605.

THE AXIS WAS NEVER EXHAUSTED -- THE CEILING WAS HIDING IT.
Wave 31 drained the block screen to 9 candidates ROM-wide and it read dry.
Every block screen this project has run had a 72-byte ceiling. Raising
--block-max to 160 returned 111 blocks, 524 candidates, 55,918 bytes. That is
wave 27's "a dry axis may mean the FLOOR, not the corpus" repeating at the other
end, and it is why this wave gained 2.63 points of bytes -- the largest jump in
recent memory -- on fewer functions than wave 31.

A DECLARATION IN unknown-globals.h SILENTLY CORRUPTED src/proc.o.
An agent added `extern int gUnknown_03001FDC;`. That symbol is
`s32 IWRAM_DATA gUnknown_03001FDC;` in src/proc.c -- upstream's own source, in
BOTH builds -- a tentative definition carrying a section attribute. src/proc.c
includes this header, so the plain extern is seen first, agbcc drops the section
attribute, and the symbol stops being placed at 0x03001FDC by `src/proc.o(.bss)`
and becomes COMMON in a stray .bss at 0x08800000. The split ROM then differed in
exactly three 4-byte pool words -- 0x08011210 and, inside proc.o itself,
0x0801C8F0 and 0x0801CB0C. Every per-function trymatch passed, because a
per-function check never links proc.o; the only symptom was the SHA.
The rule generalises: we already knew not to EDIT src/proc.c and
src/title-screen.c. DECLARING one of their globals is the same hazard, because
this header reaches them. Recorded at the declaration site.

tools/awlib.py: estimate_size counted a label-prefixed pool word as an
instruction. `_0806C99C: .4byte sym` begins with the label, so DIR_RE never saw
it and the word scored 2 bytes instead of 4. Affected functions were indexed
SHORT and trymatch slices the ROM to the indexed size, so a byte-perfect
candidate could not match at any spelling. Three functions: sub_0806C95C 82->88
(now matches, was unmatchable), sub_0808BBA4 22->24, sub_0808D680 88->90. No
promoted function was affected, so the ROM never depended on it.

Four units backed out rather than shipped, each for a reason worth keeping:
  c_080111C8.c  needs the gUnknown_03001FDC gate above; cannot be declared
  c_080114A0.c  its pool word sits at the front of a 1,412-byte data region and
                split_rodata.py can only carve an exactly-4-byte chunk
  c_08064474.c, c_0806CFC8.c  invent RAM globals at addresses upstream's lds
                does not name (0x03000600/0602, 0x03000614/0616 fall INSIDE
                gUnknown_030005FC and gUnknown_03000610). The RAM symbol table
                passes through from the untouched upstream lds, so a global
                cannot be invented at an unnamed address.

Also this wave:
- The wave-31 retype of sub_0803DF98's out-parameter to
  struct Unk02028360Pos * is CONFIRMED: its six 0x0803E callers all matched,
  and they were the first C callers that function ever had.
- A stale work draft caused a real regression. sub_08014740's declaration was
  retyped u16,u16 -> s16,s16 with the callers re-verified, but its own draft was
  left at u16 -- `conflicting types`, reading as a regression. Both spellings are
  byte-exact for that function, so the header is the only thing that could be
  wrong; draft synced to s16 and all six functions in the chain re-verified.
- 45 functions are now parked, up from 29.

+++ src/decomp/c_08025CF0.c
@@ -14,7 +14,11 @@
  * sub_08025CF0 @ 0x08025CF0
  */
 
-u8 sub_08025CF0(int index)
+/* CORRECTION, wave 32 (W32-A): `int`, not `u8`. sub_080264BC -- the first C
+ * caller either of this pair has ever had -- takes the result with a bare
+ * `adds r4, r0, #0` and no re-narrowing, and agbcc re-narrows a u8-returning
+ * callee at every call site. Byte-identical here either way. */
+int sub_08025CF0(int index)
 {
     return gUnknown_08499598[index].unk3a;
 }
