+++ src/decomp/c_08025D08.c
@@ -14,7 +14,9 @@
  * sub_08025D08 @ 0x08025D08
  */
 
-u8 sub_08025D08(int index)
+/* CORRECTION, wave 32 (W32-A): `int`, not `u8` -- see c_08025CF0.c, its twin.
+ * sub_080264BC takes this result with a bare `adds r1, r0, #0`. */
+int sub_08025D08(int index)
 {
     return gUnknown_08499598[index].unk3b;
 }
